requirements/dev.txt
django-filter
django-cors-headers==4.7.0
djangorestframework==3.16.0
django==5.2.0
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
gunicorn
whitenoise